+++ components/section-layout.tsx
@@ -2,12 +2,21 @@ import React, { ReactElement } from 'react'
 import styled from 'styled-components'
 
 const $Container = styled.div`
+  margin-top: 100px;
   margin-left: 200px;
-  position: absolute;
+  width: 100%;
+  max-width: 1000px;
+  position: relative;
 `
-const $Title = styled.div`
-  margin-top: 60px;
-  font-size: 32px;
+const $Title = styled.h1`
+  color: ${({ theme }) => theme.palette.colors.wisteria};
+  z-index: 1;
+  position: absolute;
+  padding: 0 30px;
+`
+
+const $Section = styled.div`
+  width: 100%;
 `
 interface SectionLayoutProps {
   title: string
@@ -17,7 +26,7 @@ export const SectionLayout = ({ title, children }: SectionLayoutProps): ReactEle
   return (
     <$Container>
       <$Title>{title}</$Title>
-      {children}
+      <$Section>{children}</$Section>
     </$Container>
   )
 }
