+++ package.json
@@ -26,9 +26,8 @@
     "@emotion/styled": "^11.10.5",
     "@lingui/react": "^3.14.0",
     "@mui/material": "^5.10.16",
-    "@mui/x-data-grid": "^5.17.13",
-    "@mui/x-data-grid-generator": "^5.17.14",
     "@mui/styled-engine-sc": "^5.10.16",
+    "@mui/x-data-grid": "^5.17.13",
     "easy-peasy": "^5.0.4",
     "next": "^12.1.6",
     "next-joi": "^2.2.1",
