+++ pages/index.tsx
@@ -2,7 +2,6 @@ import { createTheme, ThemeProvider } from '@mui/material/styles'
 import Head from 'next/head'
 import { ReactElement } from 'react'
 import { MainLayout } from '../components/main-layout'
-import OrdersTable from '../components/OrdersTable'
 
 const darkTheme = createTheme({
   palette: {
@@ -18,10 +17,8 @@ const Landing = (): ReactElement => {
           <title>My App</title>
           <link rel="icon" href="/favicon.ico" />
         </Head>
-
         <main>
           <MainLayout />
-          <OrdersTable />
         </main>
       </ThemeProvider>
     </div>
