+++ styles/variables.ts
@@ -15,3 +15,6 @@ export const CORAL_RED = '#ED4545'
 export const EMERALD = '#33DD77'
 export const SMALT = '#003F9E'
 export const NAVY_BLUE = '#0066FF'
+export const WISTERIA = '#c38fff'
+export const NERO = '#1e1e1e'
+export const ROBINS_EGG_BLUE = '#00dac6'
